ai/requirements.txt
Flask
requests
google-generativeai
pandas
pillow
